+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/HUD/AvatarEditorHUD/Scripts/UIToggles/FontOnToggle.cs
@@ -1,0 +1,13 @@
+﻿using UnityEngine.UI;
+using UnityEngine;
+
+public class FontOnToggle : UIToggle
+{
+    public Text targetText;
+
+    public Font onFont;
+
+    public Font offFont;
+
+    protected override void OnValueChanged(bool isOn) => targetText.font = isOn ? onFont : offFont;
+}
