+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/HUD/AvatarEditorHUD/Scripts/UIToggles/ImageColorOnToggle.cs
@@ -9,5 +9,5 @@ public class ImageColorOnToggle : UIToggle
 
     [SerializeField] private Color offColor;
 
-    protected override void OnValueChanged(bool isOn) { targetImage.color = isOn ? onColor : offColor; }
+    protected override void OnValueChanged(bool isOn) => targetImage.color = isOn ? onColor : offColor;
 }
