+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/HUD/AvatarEditorHUD/Scripts/UIToggles/TMPFontOnToggle.cs
@@ -1,0 +1,12 @@
+﻿using TMPro;
+
+public class TMPFontOnToggle : UIToggle
+{
+    public TextMeshProUGUI targetText;
+
+    public TMP_FontAsset onFont;
+
+    public TMP_FontAsset offFont;
+
+    protected override void OnValueChanged(bool isOn) => targetText.font = isOn ? onFont : offFont;
+}
